refactor: extract LayerMappingBuilder from ViewModel layout mapping
Move pipeline functions (augmentWithPushMsgActions, enrichWithCustomShiftLabels,
buildRemapOutputMap, mergeAugmentation) and all push-msg parsing into a dedicated
LayerMappingBuilder enum. ViewModel retains thin forwarding wrappers for external
callers. Reduces ViewModel+LayoutMapping by 339 lines.

14 tests cover merge invariants, remap output map, push-msg extraction, system
action labels, and the HJKL vimLabel survival invariant.
